add update api

+++ app.js
@@ -7,7 +7,7 @@ app.use(express.json());
 
 const { multer, storage } = require("./middleware/multerConfig");
 const upload = multer({ storage: storage });
-const fs = require('fs')
+const fs = require("fs");
 
 connectDatabase();
 
@@ -57,36 +57,52 @@ app.get("/blog/:id", async (req, res) => {
 });
 
 app.delete("/blog/:id", async (req, res) => {
-    const id = req.params.id
-    const blog = await Blog.findById(id)
-    const imageName = blog.image
-    fs.unlink(`storage/${imageName}` , (err) => {
-        if(err){
-            console.log(err)
-        }else{
-            console.log("File deleted successfully")
-        }
-    })
+  const id = req.params.id;
+  const blog = await Blog.findById(id);
+  const imageName = blog.image;
+  fs.unlink(`storage/${imageName}`, (err) => {
+    if (err) {
+      console.log(err);
+    } else {
+      console.log("File deleted successfully");
+    }
+  });
 
-    await Blog.findByIdAndDelete(id)
-    res.status(200).json({
-        message : "Blog deleted successfully"
-    })
-})
+  await Blog.findByIdAndDelete(id);
+  res.status(200).json({
+    message: "Blog deleted successfully",
+  });
+});
 
-app.patch("/blog/:id", async (req, res) => {
-    const id = req.params.id
-    const {title, subtitle, description} = req.body
+app.patch("/blog/:id", upload.single("image"), async (req, res) => {
+  const id = req.params.id;
+  const { title, subtitle, description } = req.body;
+  let imageName;
 
-    await Blog.findByIdAndUpdate(id, {
-        title: title,
-        subtitle: subtitle,
-        description: description
-    })
-    res.status(200).json({
-        message : "Updated successfully"
-    })
-})
+  if (req.file) {
+    imageName = req.file.filename
+    const blog = await Blog.findById(id);
+    const oldImageName = blog.image;
+
+    fs.unlink(`storage/${imageName}`, (err) => {
+      if (err) {
+        console.log(err);
+      } else {
+        console.log("File edited successfully");
+      }
+    });
+  }
+
+  await Blog.findByIdAndUpdate(id, {
+    title: title,
+    subtitle: subtitle,
+    description: description,
+    image: imageName
+  });
+  res.status(200).json({
+    message: "Updated successfully",
+  });
+});
 
 app.use(express.static("./storage"));
